Add some rudimentary test interface.
I have added a basic form with one input to submit todo items. This is
to test if the publish-subscribe design pattern.

So far the logic is simple and works the following way:

1. App renders and subscribes to an event called TODO_ADDED.
2. User submits the form with a TODO item.
3. PublishSubscribe component brodcasts an event called TODO_ADDED.
4. App hears that event and fires relevant methods.

Right now, the App doesn't know anything about the form, and the form
doesn't know anything about the App. All communication is being handled
by the PublishSubscribe component.

This will be the main design goal going forward: keep components
as independed as possible and facilitate their communication through
a pubsub component.