Add basic functionality
The app is still barebones at this point. I've just added a basic logic
of adding todos and updating the ui component.

I am still designing how the components should work. Currently, the
TodoList UI component is listening for a TODO_ADDED event.

When PublishSubscribe broadcasts that event, the UI Component re-renders
the list of todos.