Implement todo removal
The todo list now listens to a click event. When that happens, the app
fires removeTodo() method and the PublishSubscribe component brodcasts
TODO_REMOVED event.

The TodoList UI component now listen to TODO_REMOVED event too, so it
knows to re-render the list when that happens.

This commit has also introduced some minor tweaks, such as resetting
the form on submit and checking if user is trying to add an empty
item.