Refactor application state and storage.
Previously the APP_DATA in the local storage looked like this:
[
    {todoObject},
    {todoObject},
    {todoObject},
]

Now it looks like this:
[
    {projectObject},
    {projectObject},
    {projectObject},
]

Every projectObject looks like this:
{
    title: "Work Stuff",
    id: uuidv4,
    todos: [todoObject, todoObject]
}

This allows me a more flexible approach where the App's state
"remembers" the currently active project and its list of todos.

TodoList component now cares which project is currently "active" and
will render a list of todos associated with that project.