Implement interface update on todo due date change
Chaning todo's due date in one part of the app, eg. Projects --> Work,
will also now be reflected in other parts of the app, e.g.
Today, This week.

Same goes for the other way around. If the user goes to the Today's list
and updates the date of a task to tomorrow, that task will be removed
from Today's list. All parts of the app will also reflect this update.